chore(server):rename file the right way
rename file the right way

[Finishes]
